test(echo): enforce SPD precision matrix and harden Torx factors
This commit adds mathematically rigorous test coverage for the previously untouched GaussianEBM and ForcedTorxThermalizer components. During the audit, we uncovered structural vulnerabilities where the precision matrix initialization did not strictly guarantee positive-definiteness, and silent JAX dtype casting flaws existed in the thermalizer loop. These components have now been hardened to properly enforce thermodynamic and typed constraints.

* Enforced strictly lower-triangular SPD initialization in `GaussianEBM` using a softplus activation on the diagonal.
* Added missing `@jaxtyped(typechecker=beartype)` decorator to `GaussianEBM.__init__`.
* Fixed a silent dtype promotion issue in `ForcedTorxThermalizer` by strictly defining `dummy_seq` with `dtype=jnp.float32`.
* Created comprehensive unit tests in `test_ebm.py` and `test_thermalizer.py` verifying edge cases and numerical bounds.

+++ src/echo/primitives/thermalizer.py
@@ -271,7 +271,7 @@ class ForcedTorxThermalizer(eqx.Module):
             return next_state, next_state
 
         keys = jax.random.split(key, seq_len)
-        dummy_seq = jnp.zeros((seq_len, 1))
+        dummy_seq = jnp.zeros((seq_len, 1), dtype=jnp.float32)
 
         scan_seq = seq if seq is not None else dummy_seq
         scan_omega = omega_seq if omega_seq is not None else dummy_seq
